Add files via upload
The digitalization of financial procedures and systems has facilitated the increase in online payment fraud which continues to lead to loss of revenue for financial institutions. As a financial organization, Blossom bank needs to be able to identify fraudulent activities so as to increase its revenue and protect the interest of its customers. This project is aimed at building a machine learning model that will help Blossom bank identify fraudulent online payments, stop such transactions, increase its revenue and maintain customer loyalty. The dataset was gotten from the 10Alytics repository. 

During the course of this project, the first challenge I faced was understanding the features of the dataset and determining their relevance to the project. While visualizing and determining the relationships between the features, I faced the challenge of choosing the rights plots that would accurately represent the relationship between the features and give accurate insights.  The most difficult challenge was building the models and determining which would be better to use. 

At every stage there were challenges, but every challenge was an opportunity to pay attention to details, do some personal research, ask for help when I needed it, or to take a break and revisit the project with a new perspective.